% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{binnedHist}
\alias{binnedHist}
\title{Create a ggplot histogram object with binned outliers}
\usage{
binnedHist(df, var_name, bottomCode = NA, topCode = NA, col = "black",
  fill = "cornflowerblue", fill_outlier_bins = "forestgreen",
  binwidth = NULL)
}
\arguments{
\item{df}{a dataframe containing the variable to be ploted}

\item{var_name}{the variable to be plotted}

\item{bottomCode}{value below which should be binned}

\item{topCode}{value above which should be binned}

\item{col}{outline color for bars (default black)}

\item{fill}{fill color for bars (dfault blue)}

\item{fill_outlier_bins}{fill color for outlier bins}

\item{binwidth}{width of bins}
}
\value{
Returns a ggplot object
}
\description{
\code{binnedHist} creates a histogram with specified outliers
binned.
}
\details{
This creates a histogram.  Upper and/or lower outliers are grouped
together and shown in a different color to differentiate from regular 
histogram bars.
}
\examples{
binnedHist(iris,"Sepal.Length", 5, 7)

}
