% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqMd.R
\name{freqMd}
\alias{freqMd}
\title{Caclulate frequencies of variable and formats for printing in markdown}
\usage{
freqMd(...)
}
\arguments{
\item{...}{A dataframe}
}
\description{
\code{freqMd} prints a frequency table for each variable passed to it.
}
\details{
freqMd
}
