% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{freqs}
\alias{freqs}
\title{Caclulate frequencies of multiple variable}
\usage{
freqs(df, ..., plot = T, sort = T, na.rm = F)
}
\arguments{
\item{df}{A dataframe (You can also pass this program a single variable and it will silently transform it to a dataframe)}

\item{...}{a variable or variables in associated dataframe. If you pass no variables, freq will be run on all variables in dataframe}

\item{plot}{if TRUE (default) prints bar chart of results.  If FALSE, no chart.}

\item{sort}{If TRUE (default), sort output in descending order of n. If FALSE, sort output in ascending order of levels}

\item{na.rm}{if FALSE (default) NAs are included in frequency list.  If TRUE, NA are removed (but reported seperately)}
}
\value{
Null
}
\description{
\code{freqs} prints a frequency table for each variable passed to it.
}
\details{
freqs


This function prints a frequency table for variables passed to it.  In
order to fit within the tidyverse, it takes as its first argument a dataframe.
Unline \code{freq} it returns null.
}
\examples{
\dontrun{
freqs(iris, Species, Sepal.Width)
}
}
