% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFreq.R
\name{listFreq}
\alias{listFreq}
\title{Frequency list of multiple variables crossed}
\usage{
listFreq(df, ..., saveResults = F, printResults = T, sortResults = T,
  levelWarning = T, na.rm = F)
}
\arguments{
\item{df}{A dataframe.}

\item{...}{A list of variables.}

\item{saveResults}{If FALSE (default) returns original dataframe. If TRUE,
returns frequency results as a dataframe. This allows user to place this
function inside a dplyr chain}

\item{printResults}{If TRUE (default), prints results to console.  Otherwise,
if FALSE, no results are printed.}

\item{sortResults}{If TRUE (default), sort output in descending order of n.
If FALSE, sort output in ascending order of levels}

\item{levelWarning}{if TRUE (default) gives an error if the variable passed
has more than 25 levels. If}

\item{na.rm}{if TRUE (default) NAs are included in frequency list.  If FALSE,
NA are removed (but reported seperately)}
}
\value{
The original dataframe or table containing frequencies, Produces
  side-effect of printed frequencie table
}
\description{
\code{listFreq} prints a frequency table of the cross of each
variable
}
\details{
This function takes a list of variables and gives counts for each
group. For example, if variable A had two levels (1,2) and variable B had 2
levels (4,6) then listcount would give counts of cases with 1,4, 1,6, 2,4,
and 2,6.
}
\examples{
listFreq(mtcars, cyl, gear)

}
