% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Plot multiple ggplots on a single page}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{plot objects}

\item{plotlist}{or a a list containing plots}

\item{file}{unknown}

\item{cols}{number of columns}

\item{layout}{unknown}
}
\description{
\code{multiplot} prints a multiple ggplots on the same page
variable
}
\details{
This function...
}
\examples{
multiplot(p1,p2,p3,p4, cols=2)

}
