% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleFreqs-package.R
\docType{package}
\name{SimpleFreqs-package}
\alias{SimpleFreqs}
\alias{SimpleFreqs-package}
\title{SimpleFreqs: Simple Frequency Table}
\description{
This package provides a clean and easy to use frequency table. It works both on the console and in Knitted (rmarkdown and Quarto) documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://martinlbarron.github.io/SimpleFreqs/}
}

}
\keyword{internal}
