% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMeans.R
\name{summaryMeans}
\alias{summaryMeans}
\title{Produces summary statistics table}
\usage{
summaryMeans(df, ..., group_by = NA, types = c("n", "nobs", "nmiss", "sum",
  "mean", "sd", "min", "p1", "p5", "p25", "median", "p75", "p95", "p99", "max"),
  saveResults = F, printResults = T, na.rm = T)
}
\arguments{
\item{df}{a dataframe containing the variables to summarize}

\item{...}{one or more variables to summarize}

\item{group_by}{if given, statistics will be computed for each level of groupby variable}

\item{types}{types of summary statistics to computed (as vector of strings). Default is all of the following: "n", "nobs", "nmiss", "sum","mean","sd", "min", "p1", "p5", "p25", "median", "p75", "p95", "p99", "max"}

\item{saveResults}{if TRUE, dataframe of results is returned.  If FAlSE (default) original dataframe is returned}

\item{printResults}{if TRUE (default), summary table written to console.  If FALSE no results written to console.}

\item{na.rm}{whether or not na.rm==T is passsed to summarizing functions.}
}
\value{
dataframe
#'
}
\description{
Produces summary statistics table
}
\examples{

}
