% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTable.R
\name{summaryTable}
\alias{summaryTable}
\title{Construct a summary table of statistics}
\usage{
summaryTable(df, var, x, y, function_name, ...)
}
\arguments{
\item{df}{A dataframe}

\item{var}{a variable in associated dataframe.}

\item{x}{col variable}

\item{y}{row variable}

\item{function_name}{function to be used to summarize results}

\item{...}{other options to pass to function}
}
\value{
Dataframe containing table
}
\description{
prints a summary table crossing two
}
\details{
SummaryTable
}
