% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Construct frequency table}
\usage{
freq(
  df,
  var = NA,
  plot = TRUE,
  sort = TRUE,
  markdown = !interactive(),
  wt = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{df}{A data frame (optionally, you can pass a variable as the first argument.)}

\item{var}{a variable in associated data frame}

\item{plot}{if TRUE (default) prints bar chart of results.  If FALSE, no chart.}

\item{sort}{If TRUE (default), sort output in descending order of n. If FALSE, sort output in ascending order of levels}

\item{markdown}{if false, formats table for Rmarkdown/Quarto inclusion.}

\item{wt}{a variable representing weights}

\item{na.rm}{if FALSE (default) NAs are included in frequency list.  If TRUE, NA are removed (but reported separately)}
}
\value{
data frame containing frequencies.
}
\description{
Prints a frequency table for the specified variable.
}
\details{
This function constructs a frequency table for the specified variable. It expects
a data frame as the first argument and an (unquoted) variable as the second argument.
It will thus fit into a tidyverse pipeline.  Alternatively, for convienience, a
vector can be passed as first argument and the var (second) argument left blank.

The console printed table is highly customizable using session, project, or global
options.  The following options are currently implemented and can be set by calling
`options()` either interactively or by including in your .Rprofile.

- SimpleFreqs.inner_table_padding
- SimpleFreqs.table_symbol
- SimpleFreqs.row_divider_symbol
- SimpleFreqs.print_table_symbol
- SimpleFreqs.print_table_total_row
- SimpleFreqs.print_table_metadata
- SimpleFreqs.print_header_divider
- SimpleFreqs.big_mark
- SimpleFreqs.decimal_digits
}
\examples{

freq(iris, Species)

}
