% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Caclulate frequencies of variable}
\usage{
freq(df, x, saveResults = F, printResults = T, sortResults = T,
  levelWarning = T, M)
}
\arguments{
\item{df}{A dataframe}

\item{x}{a variable in associated dataframe.}

\item{saveResults}{If FALSE (default) returns original dataframe. If TRUE, returns frequency results as a dataframe. This allows user to place this function inside a dplyr chain}

\item{sortResults}{If TRUE (default), sort output in descending order of n. If FALSE, sort output in ascending order of levels}

\item{levelWarning}{if TRUE (default) gives an error if the variable passed has more than 25 levels. If}

\item{print_results}{If TRUE (default), prints results to console.  Otherwise, if FALSE, no results are printed.}
}
\value{
The original dataframe or table containing frequencies, Produces side-effect of printed frequencie table
}
\description{
\code{freq} prints a frequency table for each variable passed to it.
}
\details{
This function prints a frequency table for variables passed to it.  In
order to fit within the tidyverse, it takes as its first argument a dataframe
and returns that (unaltered) dataframe, unless the user explicilty requests other output
}
\examples{
mbfreq(iris, Species)


}
