% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFreq.R
\name{listFreq}
\alias{listFreq}
\title{Frequency list of multiple variables crossed}
\usage{
listFreq(
  df,
  ...,
  plotResults = T,
  saveResults = F,
  printResults = T,
  sortResults = F,
  levelError = T,
  na.rm = F
)
}
\arguments{
\item{df}{A dataframe.}

\item{...}{A list of variables.}

\item{plotResults}{if TRUE (default) prints bar chart of results.  If FALSE, no chart.}

\item{saveResults}{If FALSE (default) returns original dataframe. If TRUE,
returns frequency results as a dataframe. This allows user to place this
function inside a dplyr chain}

\item{printResults}{If TRUE (default), prints results to console.  Otherwise,
if FALSE, no results are printed.}

\item{sortResults}{If FALSE (default), prints results in cross order.  Otherwise,
if TRUE, sort results by frequency.}

\item{levelError}{if TRUE (default) gives an error if the variable passed
has more than 25 levels. If}

\item{na.rm}{if TRUE (default) NAs are included in frequency list.  If FALSE,
NA are removed (but reported seperately)}
}
\value{
The original dataframe or table containing frequencies, Produces
  side-effect of printed frequencie table
}
\description{
\code{listFreq} prints a frequency table of the cross of each
variable
}
\details{
listFreq

This function takes a list of variables and gives counts for each
group. For example, if variable A had two levels (1,2) and variable B had 2
levels (4,6) then listcount would give counts of cases with 1,4, 1,6, 2,4,
and 2,6.
}
\examples{
\dontrun{
listFreq(mtcars, cyl, gear)
}
}
